LC CPP

+++ CPP [2001-2500]/2485.find_the_pivot_integer [1].cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int pivotInteger(int n) {
         double sr = sqrt(n*(n+1)/2);
-        return floor(sr) == ceil(sr) ? sr : -1;
+        return floor(sr) == ceil(sr) ? (int)sr : -1;
     }
 };
 /*
@@ -11,8 +11,8 @@ public:
 # sum(1, x) = sum(x, n)
 # sum(1, x-1) = sum(x+1, n)
 # sum(1, x-1) = sum(1, n) - sum(1, x)
-# x(x-1)/2 = n(n+1)/2 - x(x+1)/2
-# x*x = n(n+1)/2
-# x = sqrt(n(n+1)/2)
-# x exists if n(n+1)/2 is a perfect square
+# x*(x-1)/2 = n*(n+1)/2 - x*(x+1)/2
+# x*x = n*(n+1)/2
+# x = sqrt(n*(n+1)/2)
+# since x is an integer, n*(n+1)/2 must be a perfect square
 */
